"Improved data validation and error handling in message appending"
This commit adds several checks in the 'appendMessageToWhatsappNumber' method of the 'firebase_conversation.py' file to ensure the required parameters (body, whatsappNumber, and sender) are not empty before proceeding. The presence of a valid conversation is also confirmed. If any of these checks fail, a HTTP 400 status and a relevant message are returned.

Furthermore, the logging setup in 'firebase_connection.py' and the 'sender' value in the document 'update_conversations.md' are modified.

A new file 'mocks.py' is introduced in 'utils' directory to carry out testing.

Lastly, 'main.py' is modified to simply return the output of 'appendMessageToWhatsappNumber' function and remove unnecessary print statement. This contributes to code clarity and reduces redundancy.