Add chaining feature for http calls and unicode support
This commit adds a new feature that allows making multiple HTTP calls in a sequence, with a predefined delay. This feature was added into a new script chainHttpCalls.py. Moreover, a function to convert the response to UTF8 was added to aid in the proper display of non-ASCII characters. Additionally, a test file that will be for automated Twilio use cases was created.