Refactor code for readability, remove unused function
This commit removes the unused function `get_conversation_by_whatsapp_number` to simplify the structure and workability of the code. It also refactors the lengthy response line in `update_conversation` for increased readability. An unnecessary newline is also removed.

In the deployment pipeline setup, deprecated unnecessary echoing and "sanitization" of environmental variable is removed, streamlining the management of environment variables for `Cloud Run` deployment. Remove deployment setup for the removed `get_conversation_by_whatsapp_number` function.