"Removed analyze_budget.py, updated .gitignore, added premium_budget.py"
Removed the "analyze_budget.py" script from the costs' folder. This script had previously been used for obtaining billing account information and authenticating requests to Google Cloud services. However, this functionality was found to be redundant and unnecessary for the current project.

Updates were made to the ".gitignore" file to include the "costs/token.json". This change ensures that leaked API tokens are prevented, increasing the security of the codebase.

Added "premium_budget.py" to handle Oauth credentials and list billing information, thereby simplifying the authentication process and making it more efficient.

These changes are crucial to improve the project's security and authentication efficiency, and overall cleaner code.