"Refactor connection classes for better readability"
Refactored a few classes in the Authentication module to enhance readability. FirebaseConnection was renamed to FirebaseSDKConnection for clarity. AbstractFirebaseConnection was introduced for encapsulation. FirebaseHTTPConnection and FirebaseHTTPAuth were added to use HTTP-based authentication. Created FirebaseConnectionFactory class for the 'factory method' design pattern, which enables the client code to create objects of a superclass and decides which ones to create.

Also, added 'requests' to requirements.txt for HTTP requests in FirebaseHTTPConnection.