Reorder environment variables and adapt password in Gmail sender
Reorganized the order of the environment variables in the `deploy-function.yml` workflow file for better readability and organization. Also made changes in `gmail_sender.py` to handle the Gmail password in a more secure way. Now the password is being adapted by breaking it into chunks of four before sending it via SMTP SSL. This adds an extra layer of security, reducing risk in case of any inadvertent exposure.