Refactor Firebase order management and improve code maintainability
The major changes incorporated entail the elimination of the premium_budget.py module and the revision of the Firebase order ID generation mechanism. The former was eliminated on account of redundancy to enhance code maintainability. The latter has been adjusted to leverage Firebase's inherent capability to generate unique IDs, hence removing the function `getNextOrderId` for order ID generation. This step strengthens reliability by significantly reducing the possibility of ID collision. These changes are also reflected in main.py to align with the updates.