Refactor CRUD operations and introduce GlobalObject
Introduced a GlobalObject class and applied refactor improvements to Create, Read, Update, and Delete (CRUD) operations for better handling of url parameters in the request. Instead of using the parameter in the header of the requests, we now have a global 'url_parameter' which is easier to manage and results in cleaner code. Also, added utility functions for request validation and standardization, providing a more robust and maintainable approach to validating request methods and extracting user id or JSON from requests. Changes are made across user_crud.py, order_crud.py, core_instantiations.py, and main.py. This approach enhances the usability of the code and makes it more flexible and scalable for the future.