Refactor Firebase connection and order creation for abstraction
This commit modifies the Firebase connection in our FirebaseOrder to use an abstract connection. This adds flexibility to our code, allowing us to easily switch between different Firebase connections if needed. Furthermore, the `createOrder` function has been updated to write data without a unique ID, solving the issue of duplicate IDs in order records. Now, the database generates a unique timestamp for each order. Also, some code has been commented out in the deploy-function file as part of a temporary fix, additional changes in the main file are for code formatting and clarity.